+++ packages/image/__tests__/android/modal-image.android.test.js
@@ -1,0 +1,16 @@
+/* eslint-env jest */
+
+import tests from "../modal-image.native.test";
+
+jest.mock("react-native", () => {
+  const reactNative = require.requireActual("react-native");
+  reactNative.Platform.OS = "android";
+  jest
+    .spyOn(reactNative.Platform, "select")
+    .mockImplementation(obj => obj.android || obj.default);
+  return reactNative;
+});
+
+describe("ModalImage test on android", () => {
+  tests();
+});